+++ src/credential_service.cpp
@@ -7,7 +7,7 @@
 #include <fstream>
 #include <credential_service.h>
 #include "curl_rest.h"
-#include "internal/credential_service_impl.h"
+#include "credential_service_impl.h"
 
 namespace {
     constexpr auto config_file_location = "token.txt";
